initial balance should be take as float

+++ src/Contexts/Adaptors/SubstrateAdaptors.tsx
@@ -144,7 +144,9 @@ export const SubstrateHomeAdaptorProvider = ({
                 homeChainConfig.tokens[0].decimals ?? homeChainConfig.decimals,
               balance: Math.max(
                 0,
-                parseInt(utils.formatUnits(balance, homeChainConfig.decimals)) -
+                parseFloat(
+                  utils.formatUnits(balance, homeChainConfig.decimals)
+                ) -
                   (homeChainConfig as SubstrateBridgeConfig)
                     .existentialDepositPlusNetworkFee
               ),
